backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.48
alembic==1.14.0
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1
yfinance==0.2.50
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.2.5
matplotlib==3.9.2
apscheduler==3.10.4
structlog==24.4.0
pytest==8.3.3
httpx==0.28.0
beautifulsoup4==4.12.3
python-multipart==0.0.18
pydantic-settings==2.6.1
